Refactor due to Siddhi upgrade and group-id changes
