+++ component/pom.xml
@@ -20,7 +20,7 @@
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <groupId>org.wso2.extension.siddhi.store.redis</groupId>
+        <groupId>io.siddhi.extension.store.redis</groupId>
         <artifactId>siddhi-store-redis-parent</artifactId>
         <version>3.1.1-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
