+++ component/src/main/java/io/siddhi/extension/store/redis/RedisIterator.java
@@ -1,0 +1,283 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.siddhi.extension.store.redis;
+
+import io.siddhi.core.table.record.RecordIterator;
+import io.siddhi.extension.store.redis.beans.StoreVariable;
+import io.siddhi.extension.store.redis.beans.StreamVariable;
+import io.siddhi.extension.store.redis.utils.RedisInstance;
+import io.siddhi.extension.store.redis.utils.RedisTableConstants;
+import io.siddhi.extension.store.redis.utils.RedisTableUtils;
+import io.siddhi.query.api.definition.Attribute;
+import redis.clients.jedis.HostAndPort;
+import redis.clients.jedis.ScanParams;
+import redis.clients.jedis.ScanResult;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+
+
+/**
+ * Redis iterator Class
+ **/
+public class RedisIterator implements RecordIterator<Object[]> {
+    private RedisInstance redisInstance;
+    private Map<String, String> resultMap = new HashMap<>();
+    private ScanResult scanResults;
+    private List<String> allResults;
+    private List<Attribute> attributes;
+    private String tableName;
+    private List scanResultsList = new ArrayList();
+    private boolean preFetched;
+    private Object[] nextValue;
+    private boolean initialTraverse;
+    private boolean isInitialTraverse = true;
+    private BasicCompareOperation query;
+    private List<String> primaryKeys;
+    private List<String> indexes;
+    private Object[] result;
+    private StoreVariable storeVariable;
+    private StreamVariable streamVariable;
+    private Long stringCursor = RedisTableConstants.REDIS_DEFAULT_CURSOR;
+    private Iterator<Object[]> iterator;
+    private List<HostAndPort> hostAndPorts;
+    private int ttl;
+    private boolean ttlOnRead;
+
+    public RedisIterator(RedisInstance redisInstance, List<Attribute> attributes, BasicCompareOperation query,
+                         String tableName, List<String> primaryKeys, List<String> indexes,
+                         List<HostAndPort> hostAndPorts, int ttl, boolean ttlOnRead) {
+        this.redisInstance = redisInstance;
+        this.attributes = attributes;
+        this.tableName = tableName;
+        this.initialTraverse = true;
+        this.query = query;
+        this.primaryKeys = primaryKeys;
+        this.indexes = indexes;
+        this.hostAndPorts = hostAndPorts;
+        this.ttl = ttl;
+        this.ttlOnRead = ttlOnRead;
+    }
+
+    @Override
+    public void close() throws IOException {
+        closeImpl();
+    }
+
+    private void closeImpl() {
+        if (resultMap != null) {
+            resultMap.clear();
+        }
+        result = null;
+        scanResultsList = null;
+        scanResults = null;
+        allResults = null;
+    }
+
+    @Override
+    public boolean hasNext() {
+        if (!this.preFetched) {
+            // check whether object is pre fetched
+            this.nextValue = this.next();
+            this.preFetched = true;
+        }
+        return nextValue != null;
+    }
+
+    @Override
+    public Object[] next() {
+        if (this.preFetched) {
+            this.preFetched = false;
+            Object[] record = this.nextValue;
+            this.nextValue = null;
+            return record;
+        }
+        List<Object[]> finalResult;
+        if (initialTraverse) {
+            finalResult = fetchResults();
+            initialTraverse = false;
+            this.iterator = finalResult.listIterator();
+            if (finalResult.isEmpty()) {
+                this.closeImpl();
+                return null;
+            }
+        }
+        if (!iterator.hasNext()) {
+            scanResultsList.clear();
+            finalResult = fetchResults();
+            this.iterator = finalResult.listIterator();
+            if (finalResult.isEmpty()) {
+                this.closeImpl();
+                return null;
+            }
+            return iterator.next();
+        } else {
+            return iterator.next();
+        }
+
+    }
+
+    @Override
+    public void remove() {
+        //Do nothing. This is a read only iterator
+    }
+
+    private List<Object[]> fetchResults() {
+        List<Object[]> resultList = new ArrayList<>();
+        if (query != null && !("true".equalsIgnoreCase((query.getStreamVariable()).getName()))) {
+            return fetchResultsWithCondition();
+        } else if (query != null && ("true".equalsIgnoreCase((query.getStreamVariable()).getName()))) {
+            return fetchAllResults();
+        }
+        return resultList;
+    }
+
+    private List<Object[]> fetchResultsWithCondition() {
+        List<Object[]> resultList = new ArrayList<>();
+        storeVariable = query.getStoreVariable();
+        streamVariable = query.getStreamVariable();
+        if (indexes.contains(storeVariable.getName())) {
+            return fetchResultsByIndexedValue();
+        } else if (primaryKeys.contains(storeVariable.getName())) {
+            return fetchResultsByPrimaryKey();
+        }
+        return resultList;
+    }
+
+    private List<Object[]> fetchResultsByPrimaryKey() {
+        List<Object[]> resultList = new ArrayList<>();
+        if (isInitialTraverse) {
+            String rowKey = tableName + ":" + streamVariable.getName();
+            resultMap = redisInstance.hgetAll(tableName + ":" + streamVariable.getName());
+            if (resultMap != null && !resultMap.isEmpty()) {
+                result = resultsGenerator(resultMap);
+                resultList.add(result);
+                isInitialTraverse = false;
+                if (ttlOnRead) {
+                  RedisTableUtils.setExpire(redisInstance, tableName, indexes, rowKey, ttl, resultMap);
+                }
+            }
+        }
+        return resultList;
+    }
+
+    private List<Object[]> fetchResultsByIndexedValue() {
+        List<Object[]> resultList = new ArrayList<>();
+        ScanParams scanParams = new ScanParams();
+        scanParams.count(RedisTableConstants.REDIS_BATCH_SIZE);
+        if (isInitialTraverse) {
+            isInitialTraverse = false;
+            scanResultsList = fetchBatchOfResultsFromIndexTable(scanParams);
+            if (scanResultsList.isEmpty()) {
+                return resultList;
+            }
+        }
+        if (scanResultsList.isEmpty() && stringCursor > 0) {
+            scanResultsList = fetchBatchOfResultsFromIndexTable(scanParams);
+            if (scanResultsList.isEmpty()) {
+                return resultList;
+            }
+        } else if (scanResultsList.isEmpty() && stringCursor == 0) {
+            return resultList;
+        }
+        scanResultsList.forEach(scanResult -> {
+            resultMap = redisInstance.hgetAll(scanResult.toString());
+            if (resultMap != null) {
+                result = resultsGenerator(resultMap);
+                resultList.add(result);
+                if (ttlOnRead) {
+                  RedisTableUtils.setExpire(redisInstance, tableName, indexes, scanResult.toString(), ttl, resultMap);
+                }
+            }
+        });
+        return resultList;
+    }
+
+    private List fetchBatchOfResultsFromIndexTable(ScanParams scanParams) {
+        scanResults = redisInstance.sscan(tableName + ":" + storeVariable.getName() + ":"
+                + streamVariable.getName(), String.valueOf(stringCursor), scanParams);
+        stringCursor = Long.valueOf(scanResults.getStringCursor());
+        return scanResults.getResult();
+    }
+
+    private List<Object[]> fetchAllResults() {
+        List<Object[]> resultList = new ArrayList<>();
+        if (isInitialTraverse) {
+            isInitialTraverse = false;
+            this.scanResultsList = fetchBatchOfResultsFromTable();
+            if (scanResultsList.isEmpty()) {
+                return resultList;
+            }
+        }
+        scanResultsList.forEach(e -> {
+            String type = redisInstance.type(e.toString());
+            if (type.equalsIgnoreCase("hash")) {
+                resultMap = redisInstance.hgetAll(e.toString());
+                if (resultMap != null) {
+                    result = resultsGenerator(resultMap);
+                    resultList.add(result);
+                    if (ttlOnRead) {
+                        RedisTableUtils.setExpire(redisInstance, tableName, indexes, e.toString(), ttl, resultMap);
+                    }
+                }
+            }
+        });
+        return resultList;
+    }
+
+    private List fetchBatchOfResultsFromTable() {
+        ScanParams scanParams = new ScanParams();
+        scanParams.match(tableName + ":*");
+        scanParams.count(RedisTableConstants.REDIS_BATCH_SIZE);
+        allResults = redisInstance.scan(hostAndPorts, scanParams);
+        return allResults;
+    }
+
+    private Object[] resultsGenerator(Map<String, String> resultMap) {
+        Object[] generatedResult = new Object[attributes.size()];
+        int i = 0;
+        for (Attribute attribute : this.attributes) {
+            switch (attribute.getType()) {
+                case BOOL:
+                    generatedResult[i++] = Boolean.parseBoolean(resultMap.get(attribute.getName()));
+                    break;
+                case INT:
+                    generatedResult[i++] = Integer.parseInt(resultMap.get(attribute.getName()));
+                    break;
+                case LONG:
+                    generatedResult[i++] = Long.parseLong(resultMap.get(attribute.getName()));
+                    break;
+                case FLOAT:
+                    generatedResult[i++] = Float.parseFloat(resultMap.get(attribute.getName()));
+                    break;
+                case DOUBLE:
+                    generatedResult[i++] = Double.parseDouble(resultMap.get(attribute.getName()));
+                    break;
+                default:
+                    generatedResult[i++] = resultMap.get(attribute.getName());
+                    break;
+            }
+        }
+        return generatedResult;
+    }
+}
