+++ component/src/main/java/io/siddhi/extension/store/redis/utils/RedisTableConstants.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.siddhi.extension.store.redis.utils;
+
+/**
+ * This class contains the constance which used in this store
+ **/
+public class RedisTableConstants {
+    private RedisTableConstants() {
+        //preventing initialization
+    }
+
+    public static final String ANNOTATION_ELEMENT_HOST = "host";
+    public static final String ANNOTATION_ELEMENT_PORT = "port";
+    public static final String ANNOTATION_ELEMENT_PASSWORD = "password";
+    public static final String ANNOTATION_ELEMENT_TABLE_NAME = "table.name";
+    public static final String ANNOTATION_ELEMENT_CLUSTER_MODE = "cluster.mode";
+    public static final String ANNOTATION_ELEMENT_NODES = "nodes";
+    public static final String ANNOTATION_ELEMENT_TTL_SECS = "ttl.seconds";
+    public static final String ANNOTATION_ELEMENT_TTL_ON_UPDATE = "ttl.on.update";
+    public static final String ANNOTATION_ELEMENT_TTL_ON_READ = "ttl.on.read";
+    public static final String DEFAULT_HOST = "localhost";
+    public static final int DEFAULT_PORT = 6379;
+    public static final int REDIS_BATCH_SIZE = 1000;
+    public static final Long REDIS_DEFAULT_CURSOR = 0L;
+}
