+++ component/src/test/resources/testng.xml
@@ -22,13 +22,13 @@
 <suite name="Redis-Event-Table-Test-Suite">
     <test name="redis-tests" preserve-order="true" parallel="false">
         <classes>
-            <class name="org.wso2.extension.siddhi.store.redis.test.RedisTestUtils"/>
-            <class name="org.wso2.extension.siddhi.store.redis.test.DefineRedisTableTestCase"/>
-            <class name="org.wso2.extension.siddhi.store.redis.test.DeleteFromRedisTableTestCase"/>
-            <class name="org.wso2.extension.siddhi.store.redis.test.InsertIntoRedisTableTestCase"/>
-            <class name="org.wso2.extension.siddhi.store.redis.test.ReadFromRedisTableTestCase"/>
-            <class name="org.wso2.extension.siddhi.store.redis.test.UpdateRedisTableTestCase"/>
-            <class name="org.wso2.extension.siddhi.store.redis.test.UpdateOrInsertRedisTableTestCase"/>
+            <class name="io.siddhi.extension.store.redis.test.RedisTestUtils"/>
+            <class name="io.siddhi.extension.store.redis.test.DefineRedisTableTestCase"/>
+            <class name="io.siddhi.extension.store.redis.test.DeleteFromRedisTableTestCase"/>
+            <class name="io.siddhi.extension.store.redis.test.InsertIntoRedisTableTestCase"/>
+            <class name="io.siddhi.extension.store.redis.test.ReadFromRedisTableTestCase"/>
+            <class name="io.siddhi.extension.store.redis.test.UpdateRedisTableTestCase"/>
+            <class name="io.siddhi.extension.store.redis.test.UpdateOrInsertRedisTableTestCase"/>
         </classes>
     </test>
 </suite>
