+++ mkdocs.yml
@@ -1,13 +1,13 @@
 site_name: Siddhi Store Redis
 site_description: Siddhi Store Redis Extension
 site_author: WSO2
-site_url: https://wso2-extensions.github.io/siddhi-store-redis/
+site_url: https://siddhi-io.github.io/siddhi-store-redis/
 extra_css:
 - stylesheets/extra.css
 extra_javascript:
 - javascripts/extra.js
 repo_name: Siddhi-Store-Redis
-repo_url: https://github.com/wso2-extensions/siddhi-store-redis
+repo_url: https://github.com/siddhi-io/siddhi-store-redis
 copyright: Copyright &copy; 2011 - 2018 WSO2
 theme:
   name: material
