+++ pom.xml
@@ -24,7 +24,7 @@
         <version>5</version>
     </parent>
 
-    <groupId>org.wso2.extension.siddhi.store.redis</groupId>
+    <groupId>io.siddhi.extension.store.redis</groupId>
     <artifactId>siddhi-store-redis-parent</artifactId>
     <version>3.1.1-SNAPSHOT</version>
     <packaging>pom</packaging>
@@ -130,14 +130,13 @@
     </profiles>
 
     <properties>
-        <siddhi.version>5.0.0</siddhi.version>
+        <siddhi.version>5.1.1-SNAPSHOT</siddhi.version>
         <siddhi.version.range>[5.0.0,6.0.0)</siddhi.version.range>
         <wso2.slf4j.version>1.5.10.wso2v1</wso2.slf4j.version>
         <testng.version>6.8</testng.version>
         <commons.logging.version>1.1.1</commons.logging.version>
         <redis.version>2.9.0</redis.version>
-        <jacoco.version>0.7.9</jacoco.version>
-        <jacoco.ant.version>0.7.9</jacoco.ant.version>
+        <jacoco.version>0.8.4</jacoco.version>
         <awaitility.version>3.1.0</awaitility.version>
         <redis.import.range>[2.8.0 3.0.0]</redis.import.range>
         <apache.commons.pool.version>2.4.0</apache.commons.pool.version>
@@ -145,9 +144,9 @@
     </properties>
 
     <scm>
-        <connection>scm:git:https://github.com/wso2-extensions/siddhi-store-redis.git</connection>
-        <url>https://github.com/wso2-extensions/siddhi-store-redis.git</url>
-        <developerConnection>scm:git:https://github.com/wso2-extensions/siddhi-store-redis.git
+        <connection>scm:git:https://github.com/siddhi-io/siddhi-store-redis.git</connection>
+        <url>https://github.com/siddhi-io/siddhi-store-redis.git</url>
+        <developerConnection>scm:git:https://github.com/siddhi-io/siddhi-store-redis.git
         </developerConnection>
         <tag>HEAD</tag>
     </scm>
@@ -170,15 +169,9 @@
                 <version>${siddhi.version}</version>
             </dependency>
             <dependency>
-                <groupId>org.wso2.extension.siddhi.store.redis</groupId>
+                <groupId>io.siddhi.extension.store.redis</groupId>
                 <artifactId>siddhi-store-redis</artifactId>
                 <version>${project.version}</version>
-                <scope>test</scope>
-            </dependency>
-            <dependency>
-                <groupId>org.awaitility</groupId>
-                <artifactId>awaitility</artifactId>
-                <version>${awaitility.version}</version>
                 <scope>test</scope>
             </dependency>
             <dependency>
@@ -238,6 +231,7 @@
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
+                <version>${jacoco.version}</version>
                 <executions>
                     <execution>
                         <id>jacoco-initialize</id>
@@ -272,7 +266,7 @@
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Export-Package>
                             org.apache.commons.pool2.*;version="${apache.commons.pool.version}",
-                            org.wso2.extension.siddhi.store.redis.*;version="${project.version}"
+                            io.siddhi.extension.store.redis.*;version="${project.version}"
                         </Export-Package>
                         <Private-Package>
                             org.apache.commons.pool2.*;version="${apache.commons.pool.version}"
